+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Features
 
 - You can change the ZX Spectrum 48 ROM
+- New text editor for text file types
+- New binary editor for viewing binary file content
+- New image viewer to display bitmap images
 
 ## 0.53.2